Refactor USB connection management so the logic is shared
ClockManager and BootloaderConnection for the programmer are both
subclasses of Connection.
